Added 'toEmail' option
This option is relevant, if you actually want to hide the original To header instead of directly forwarding the original email.
For me this is necessary, since the original 'To:' email should not be included in the forwarded email, to make the process as transparent to the receiver as possible. 

Note: For me, this option does only make sense, if I also use the "fromEmail" option. Otherwise the "From:" header is set to the value of the original "To:" header, which would include the original email destination, too.